Запилил ответные удары

+++ app/models/armor.go
@@ -2,10 +2,10 @@ package models
 
 type Armor struct {
 	style string
-	protectionValue int
+	protectionValue float32
 }
 
-func NewArmor(style string, protectionValue int) *Armor {
+func NewArmor(style string, protectionValue float32) *Armor {
 	return &Armor{
 		style:   style,
 		protectionValue: protectionValue,
@@ -16,6 +16,6 @@ func (armor *Armor) GetStyle() string {
 	return armor.style
 }
 
-func (armor *Armor) GetProtectionValue() int {
+func (armor *Armor) GetProtectionValue() float32 {
 	return armor.protectionValue
 }
